# DigiMediaStore/Controllers/UserController.cs
using DigiMediaStore.BusinessLogic.Interfaces;
using DigiMediaStore.DataAccess.Models;
using DigiMediaStore.Models;
using Microsoft.AspNetCore.Mvc;

namespace DigiMediaStore.Controllers;

[Route("api/[controller]")]
[ApiController]
public class UserController : ControllerBase
{
    private IUserService _userService;
    
    public UserController(IUserService userService)
    {
        _userService = userService;
    }

    [HttpGet]
    public async Task<IActionResult> GetAll()
    {
        return Ok(await _userService.GetAll());
    }

    [HttpGet("{id}")]
    public async Task<IActionResult> GetById(int id)
    {
        return Ok(await _userService.GetById(id));
    }

    [HttpPost]
    public async Task<IActionResult> Add(User user)
    {
        await _userService.Create(user);
        return Ok();
    }

    [HttpPut]
    public async Task<IActionResult> Update(User user)
    {
        await _userService.Update(user);
        return Ok();
    }

    [HttpDelete]
    public async Task<IActionResult> Delete(int id)
    {
        await _userService.Delete(id);
        return Ok();
    }

    [HttpPost("test-api")]
    public IActionResult TestApi([FromBody] ApiInfo apiInfo)
    {
        // Этот endpoint нужен только для отображения ApiInfo в Swagger
        return Ok(apiInfo);
    }
}